Add unit types
These will help to avoid unit errors by having specific types
for units which requires to explicitly convert to other units
or to access the raw value when doing computations. This way
there can't be any implicit conversion between units and the
unit can be guaranteed by the type system.